feat: Reorganised, better toolbar styles
