+++ starter-kits/nextjs-starter-kit/components/TextEditor/Toolbar.module.css
@@ -1,5 +1,6 @@
 .toolbar {
   display: flex;
+  align-items: center;
   gap: var(--space-2);
 }
 
@@ -10,4 +11,11 @@
 .toolbarButton svg {
   height: 20px;
   width: 20px;
+}
+
+.toolbarSeparator {
+  width: 1px;
+  height: 65%;
+  background: var(--color-border);
+  margin: 0 var(--space-2);
 }
