+++ starter-kits/nextjs-starter-kit/components/TextEditor/ToolbarItems/Headings.module.css
@@ -1,0 +1,3 @@
+.headingsSelect {
+  width: 130px;
+}
